QR_to_Website/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
qrcode
pyqrcode
Pillow